Adding TaskHandler abstraction layer.
