+++ core/src/main/java/liquid/core/controller/BaseController.java
@@ -1,6 +1,5 @@
 package liquid.core.controller;
 
-import org.springframework.stereotype.Component;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 
@@ -9,7 +8,6 @@ import org.springframework.validation.FieldError;
  * Date: 10/16/13
  * Time: 9:24 PM
  */
-@Component
 public abstract class BaseController {
     protected static final int size = 20;
