+++ process/process-api/src/main/java/liquid/process/controller/BaseTaskController.java
@@ -1,10 +1,9 @@
 package liquid.process.controller;
 
+import liquid.core.controller.BaseController;
 import liquid.process.domain.Task;
 import liquid.process.service.TaskService;
-import liquid.core.controller.BaseController;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -15,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestParam;
  * Date: 10/4/13
  * Time: 5:42 PM
  */
-@Component
+@Deprecated
 public abstract class BaseTaskController extends BaseController {
     @Autowired
     protected TaskService taskService;
