+++ process/process-web/src/main/java/liquid/process/controller/AllocateContainersController.java
@@ -1,0 +1,10 @@
+package liquid.process.controller;
+
+import org.springframework.stereotype.Controller;
+
+/**
+ * Created by Tao Ma on 4/16/15.
+ */
+@Controller
+public class AllocateContainersController extends AbstractTaskController {
+}
