+++ process/process-web/src/main/java/liquid/process/controller/FeedDistyPriceController.java
@@ -1,12 +1,12 @@
-package liquid.controller;
+package liquid.process.controller;
 
-import liquid.process.model.Disty;
 import liquid.core.model.Alert;
 import liquid.order.domain.OrderEntity;
 import liquid.order.service.OrderService;
+import liquid.process.handler.FeedDistyPriceHandler;
+import liquid.process.model.Disty;
 import liquid.process.service.TaskService;
 import liquid.security.SecurityContext;
-import liquid.core.controller.BaseController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,24 +14,21 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 import java.util.Date;
 
 /**
- *  
- * User: tao
- * Date: 10/24/13
- * Time: 9:47 PM
+ * Created by Tao Ma on 4/15/15.
  */
 @Controller
-@RequestMapping("/dp")
-public class OrderDistyController extends BaseController {
-    private static final Logger logger = LoggerFactory.getLogger(OrderDistyController.class);
+public class FeedDistyPriceController extends AbstractTaskController {
+
+    private static final Logger logger = LoggerFactory.getLogger(FeedDistyPriceController.class);
 
     @Autowired
     private TaskService taskService;
@@ -39,31 +36,17 @@ public class OrderDistyController extends BaseController {
     @Autowired
     private OrderService orderService;
 
-    @RequestMapping(method = RequestMethod.GET)
-    public String init(@RequestParam(value = "t") String taskId, Model model) {
-        logger.debug("taskId: {}", taskId);
-
-        long orderId = taskService.getOrderIdByTaskId(taskId);
-        OrderEntity order = orderService.find(orderId);
-        Disty disty = new Disty();
-        disty.setDistyCny(order.getDistyCny());
-        disty.setDistyUsd(order.getDistyUsd());
-        model.addAttribute("disty", disty);
-        model.addAttribute("task", taskService.getTask(taskId));
-        return "order/disty_price";
-    }
-
-    @RequestMapping(method = RequestMethod.POST)
-    public String feed(@RequestParam(value = "t") String taskId,
+    @RequestMapping(method = RequestMethod.POST, params = "definitionKey=" + FeedDistyPriceHandler.TASK_DEFINITION_KEY)
+    public String feed(@PathVariable String taskId,
                        @Valid @ModelAttribute("disty") Disty disty, BindingResult bindingResult,
                        Model model, RedirectAttributes redirectAttributes) {
         logger.debug("taskId: {}", taskId);
 
         if (bindingResult.hasErrors()) {
-            return "order/disty_price";
+            return locateTemplate(FeedDistyPriceHandler.TASK_DEFINITION_KEY);
         }
 
-        long orderId = taskService.getOrderIdByTaskId(taskId);
+        Long orderId = taskService.getOrderIdByTaskId(taskId);
         OrderEntity order = orderService.find(orderId);
         order.setDistyCny(disty.getDistyCny());
         order.setDistyUsd(disty.getDistyUsd());
@@ -72,6 +55,6 @@ public class OrderDistyController extends BaseController {
         orderService.save(order);
 
         redirectAttributes.addFlashAttribute("alert", new Alert("save.success"));
-        return "redirect:/dp?t=" + taskId;
+        return computeRedirect(taskId);
     }
 }
