+++ process/process-web/src/main/java/liquid/process/handler/FeedContainerNoHandler.java
@@ -1,14 +1,16 @@
-package liquid.process.service;
+package liquid.process.handler;
 
 import liquid.order.domain.OrderEntity;
 import liquid.order.service.OrderService;
 import liquid.process.NotCompletedException;
+import liquid.process.domain.Task;
 import liquid.transport.domain.ShipmentEntity;
 import liquid.transport.domain.ShippingContainerEntity;
 import liquid.transport.service.ShipmentService;
 import liquid.transport.service.ShippingContainerService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
+import org.springframework.ui.Model;
 
 import java.util.Collection;
 import java.util.Map;
@@ -16,8 +18,8 @@ import java.util.Map;
 /**
  * Created by redbrick9 on 7/22/14.
  */
-@Service("feedContainerNo")
-public class FeedContainerNoTask extends AbstractTaskProxy {
+@Component
+public class FeedContainerNoHandler extends AbstractTaskHandler {
     @Autowired
     private OrderService orderService;
 
@@ -47,5 +49,25 @@ public class FeedContainerNoTask extends AbstractTaskProxy {
         if (allocatedContainerQuantity != order.getContainerQty()) {
             throw new NotCompletedException("container.allocation.is.not.completed");
         }
+    }
+
+    @Override
+    public String getDefinitionKey() {
+        return null;
+    }
+
+    @Override
+    public boolean isRedirect() {
+        return true;
+    }
+
+    @Override
+    public void init(Task task, Model model) {
+
+    }
+
+    @Override
+    public void claim(Task task) {
+
     }
 }
