+++ process/process-web/src/main/java/liquid/process/handler/PlanRouteHandler.java
@@ -1,0 +1,12 @@
+package liquid.process.handler;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * Created by Tao Ma on 4/5/15.
+ */
+@Deprecated
+@Component
+public class PlanRouteHandler extends PlanShipmentHandler {
+    // FIXME - Remove this class till the old processes are finished.
+}
