+++ process/process-web/src/main/java/liquid/process/handler/SendInvoicingHandler.java
@@ -1,17 +1,19 @@
-package liquid.process.service;
+package liquid.process.handler;
 
 import liquid.accounting.domain.ReceivableSummaryEntity;
 import liquid.accounting.service.ReceivableSummaryService;
+import liquid.process.domain.Task;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
+import org.springframework.ui.Model;
 
 import java.util.Map;
 
 /**
  * Created by redbrick9 on 6/7/14.
  */
-@Service("sendInvoicing")
-public class SendInvoiceTask extends AbstractTaskProxy {
+@Component
+public class SendInvoicingHandler extends AbstractTaskHandler {
     @Autowired
     private ReceivableSummaryService receivableSummaryService;
 
@@ -20,5 +22,25 @@ public class SendInvoiceTask extends AbstractTaskProxy {
         Long orderId = taskService.getOrderIdByTaskId(taskId);
         ReceivableSummaryEntity receivableSummaryEntity = receivableSummaryService.findByOrderId(orderId);
         variableMap.put("salesPrice", "CNY: " + receivableSummaryEntity.getCny() + "; USD: " + receivableSummaryEntity.getUsd());
+    }
+
+    @Override
+    public String getDefinitionKey() {
+        return null;
+    }
+
+    @Override
+    public boolean isRedirect() {
+        return true;
+    }
+
+    @Override
+    public void init(Task task, Model model) {
+
+    }
+
+    @Override
+    public void claim(Task task) {
+
     }
 }
