--- process/process-web/src/main/java/liquid/process/service/PlanRouteTask.java
@@ -1,12 +1,0 @@
-package liquid.process.service;
-
-import org.springframework.stereotype.Service;
-
-/**
- * Created by Tao Ma on 4/5/15.
- */
-@Deprecated
-@Service("planRoute")
-public class PlanRouteTask extends PlanShipmentTask {
-    // FIXME - Remove this class till the old processes are finished.
-}
