+++ web-console/src/main/webapp/WEB-INF/views/task/sendTruck/init.html
@@ -74,8 +74,8 @@
 
         <div class="row">
             <div class="col-xs-3 pull-right">
-                <input type="hidden" th:field="*{definitionKey}" class="form-control"/>
-                <input type="hidden" th:field="*{orderId}" class="form-control"/>
+                <input type="hidden" th:field="*{orderId}"/>
+                <input type="hidden" id="definitionKey" name="definitionKey" th:value="${task.definitionKey}"/>
                 <input type="submit" value="Fill In" class="btn btn-success btn-block btn-sm" th:value="#{save}"/>
             </div>
         </div>
